pkg/config: extract filter strict validators into compiler_validate_strict_filter.go (#4405)
Pure code-motion split of the 6,997-LOC compiler_validate_strict.go
(#4405). Moves the firewall-filter and protocol/DSCP-resolver strict commit-time validators
(validateXxxStrict) and their domain-local helpers verbatim into
compiler_validate_strict_filter.go (package config unchanged).

No validator logic, signature, or name changed: declarations are moved
byte-for-byte and only the per-file import set is recomputed. Verified by
the sorted top-level declaration-name set staying identical across the
split, and by go build ./... plus go test ./pkg/config/... staying green
at this step. Repoints the #3393 Go<->Rust filter-protocol parity test at the
new file (its source anchor for filterProtocolResolvable is path-based).
